wip: statements modal, refactor and cleanup
