+++ src/components/Billing/AccountStatements.tsx
@@ -1,9 +1,12 @@
 import { useMemo, useCallback } from 'react'
-import { Title, Button, Flex, Group, Loader, Modal, createStyles } from '@mantine/core'
+import { Title } from '@mantine/core'
 import CustomTable from 'components/common/CustomTable'
 import { useDisclosure } from '@mantine/hooks'
 import useAccount from 'hooks/useAccount'
 import { Deposit, CampaignFundsActive, CampaignRefunds, OperationEntry, StatementData } from 'types'
+import { ADEX_COMPANY_DETAILS } from 'constants/adexCompanyDetatils'
+import { BillingDetailsModal } from './BillingDetailsModal'
+import { StatementsPDF } from './BillingPDF'
 
 const columnTitles = ['Date of issue', 'Token']
 
@@ -37,87 +40,10 @@ const toOpEntry = (
   }
 }
 
-const useStyles = createStyles((theme) => ({
-  wrapper: {
-    border: '1px solid',
-    borderRadius: theme.radius.sm,
-    borderColor: theme.colors.decorativeBorders[theme.fn.primaryShade()],
-    padding: theme.spacing.lg
-  },
-  header: {
-    backgroundColor: theme.colors.lightBackground[theme.fn.primaryShade()],
-    padding: theme.spacing.xl
-  },
-  title: {
-    fontSize: theme.fontSizes.xl,
-    fontWeight: theme.other.fontWeights.bold
-  },
-  close: {
-    color: theme.colors.mainText[theme.fn.primaryShade()]
-  },
-  printable: {
-    [theme.other.media.print]: {
-      // NOTE: it's not fixed/absolute to body but modal.inner
-      overflow: 'visible',
-      position: 'absolute',
-      top: 0,
-      left: 0,
-      bottom: 0,
-      width: '100%',
-      padding: theme.spacing.xl
-    }
-  }
-}))
-
-type ModalProps = {
-  statement: StatementData
-  opened: boolean
-  close: () => void
-}
-
-const StatementModal = ({ statement, opened, close }: ModalProps) => {
-  const { classes } = useStyles()
-  return (
-    <Modal
-      title="Invoice"
-      size="xl"
-      opened={opened}
-      onClose={close}
-      centered
-      radius="sm"
-      classNames={{
-        header: classes.header,
-        title: classes.title,
-        close: classes.close
-      }}
-    >
-      <div>
-        {!statement ? (
-          <Flex justify="center" align="center" h="60vh">
-            <Loader size="xl" />
-          </Flex>
-        ) : (
-          <>
-            <Group position="right">
-              <Button mt="md" mb="md" onClick={() => window.print()}>
-                Print
-              </Button>
-            </Group>
-
-            <div className={classes.wrapper}>
-              <div id="printable" className={classes.printable} />
-            </div>
-          </>
-        )}
-      </div>
-    </Modal>
-  )
-}
-
 const AccountStatements = () => {
   const [opened, { open, close }] = useDisclosure(false)
   const {
-    adexAccount: { fundsDeposited, fundsOnCampaigns, refundsFromCampaigns }
+    adexAccount: { address, billingDetails, fundsDeposited, fundsOnCampaigns, refundsFromCampaigns }
   } = useAccount()
 
   const statements = useMemo(() => {
@@ -208,10 +134,17 @@ const AccountStatements = () => {
     return <Title order={4}>No AccountStatements found.</Title>
   }
 
+  const statement = useMemo(() => statements[0], [statements])
+
   return (
     <>
-      {/* Temporary disabled */}
-      <StatementModal statement={statements[0]} opened={opened} close={close} />
+      <BillingDetailsModal title="Statement" loading={!statement} opened={opened} close={close}>
+        <StatementsPDF
+          statement={statement}
+          seller={ADEX_COMPANY_DETAILS}
+          buyer={{ ...billingDetails, ethAddress: address }}
+        />
+      </BillingDetailsModal>
       <CustomTable headings={columnTitles} elements={elements} onPreview={onPreview} />
     </>
   )
