+++ src/components/Billing/InvoicesModal.tsx
@@ -1,4 +1,3 @@
-import { Button, Flex, Group, Loader, Modal, createStyles } from '@mantine/core'
 import { InvoicesPDF } from 'components/Billing/BillingPDF'
 import { ADEX_COMPANY_DETAILS } from 'constants/adexCompanyDetatils'
 import useAccount from 'hooks/useAccount'
@@ -6,38 +5,7 @@ import useCampaignAnalytics from 'hooks/useCampaignAnalytics'
 import useCampaignsData from 'hooks/useCampaignsData'
 import { useEffect, useMemo, useState } from 'react'
 import { AnalyticsPeriod, BaseAnalyticsData, IInvoiceDetails } from 'types'
-
-const useStyles = createStyles((theme) => ({
-  wrapper: {
-    border: '1px solid',
-    borderRadius: theme.radius.sm,
-    borderColor: theme.colors.decorativeBorders[theme.fn.primaryShade()],
-    padding: theme.spacing.lg
-  },
-  header: {
-    backgroundColor: theme.colors.lightBackground[theme.fn.primaryShade()],
-    padding: theme.spacing.xl
-  },
-  title: {
-    fontSize: theme.fontSizes.xl,
-    fontWeight: theme.other.fontWeights.bold
-  },
-  close: {
-    color: theme.colors.mainText[theme.fn.primaryShade()]
-  },
-  printable: {
-    [theme.other.media.print]: {
-      // NOTE: it's not fixed/absolute to body but modal.inner
-      overflow: 'visible',
-      position: 'absolute',
-      top: 0,
-      left: 0,
-      bottom: 0,
-      width: '100%',
-      padding: theme.spacing.xl
-    }
-  }
-}))
+import { BillingDetailsModal } from './BillingDetailsModal'
 
 type PrintModalProps = {
   campaignId: string
@@ -116,45 +84,17 @@ export const InvoicesModal = ({ campaignId, opened, close }: PrintModalProps) =>
     }
   }, [address, billingDetails, campaign?.id, campaignMappedAnalytics, currencyName])
 
-  const { classes } = useStyles()
   return (
-    <Modal
+    <BillingDetailsModal
       title="Invoice"
-      size="xl"
+      loading={!analyticsKey || !campaignMappedAnalytics}
       opened={opened}
-      onClose={close}
-      centered
-      radius="sm"
-      classNames={{
-        header: classes.header,
-        title: classes.title,
-        close: classes.close
-      }}
+      close={close}
     >
-      <div>
-        {!analyticsKey || !campaignMappedAnalytics ? (
-          <Flex justify="center" align="center" h="60vh">
-            <Loader size="xl" />
-          </Flex>
-        ) : (
-          <>
-            <Group position="right">
-              <Button mt="md" mb="md" onClick={() => window.print()}>
-                Print
-              </Button>
-            </Group>
-
-            <div className={classes.wrapper}>
-              <div id="printable" className={classes.printable}>
-                <InvoicesPDF
-                  invoiceDetails={elements}
-                  placement={campaign?.targetingInput.inputs.placements.in[0] || 'site'}
-                />
-              </div>
-            </div>
-          </>
-        )}
-      </div>
-    </Modal>
+      <InvoicesPDF
+        invoiceDetails={elements}
+        placement={campaign?.targetingInput.inputs.placements.in[0] || 'site'}
+      />
+    </BillingDetailsModal>
   )
 }
