+++ src/components/Billing/mockedData.ts
@@ -1,5 +1,3 @@
-import { IStatements } from 'types'
-
 // const invoiceElements: IInvoices[] = [
 const invoiceElements = [
   {
@@ -603,258 +601,6 @@ const invoiceElements = [
     amountSpent: '3000.00 DAI'
   }
 ]
-const statementElements: IStatements[] = [
-  {
-    id: 1,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 2,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 3,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 4,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 5,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 6,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 7,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 8,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 9,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 10,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 11,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 12,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 13,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 14,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 15,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 16,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 17,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 18,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 19,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 20,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 21,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 22,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 23,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 24,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 25,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 26,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 27,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 28,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 29,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 30,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 31,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 32,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 33,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 34,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 35,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 36,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 37,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 38,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 39,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 40,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 41,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 42,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 43,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 44,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 45,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 46,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 47,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 48,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 49,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  },
-  {
-    id: 50,
-    documentName: 'Statement March 2023',
-    dateOfIssue: '31.03.2023'
-  }
-]
 // const invoiceElements: IInvoices[] | null = null
 // const statementElements: IStatements[] | null = null
 
@@ -969,4 +715,4 @@ const invoiceDetails = {
   vatPercentageInUSD: 10
 }
 
-export { invoiceElements, statementElements, invoiceDetails }
+export { invoiceElements, invoiceDetails }
