+++ src/types/customTable.ts
@@ -1,10 +1,10 @@
 import { PropsWithChildren } from 'react'
-import { IStatements, IInvoices } from 'types'
+import { IInvoices } from 'types'
 
 export interface ICustomTableProps extends PropsWithChildren {
   background?: boolean
   headings: string[]
-  elements: IInvoices[] | IStatements[] | any[]
+  elements: IInvoices[] | any[]
   onPreview?: (e: any) => any
   onAnalytics?: (e: any) => any
   onDuplicate?: (e: any) => any
